+++ src/watermelonSidebar.ts
@@ -117,6 +117,12 @@ export default class WatermelonSidebar implements vscode.WebviewViewProvider {
           });
           break;
         }
+        default:{
+          this.sendMessage({
+            command: "",
+          });
+        }
+
       }
     });
   }